VanillaLabs Site
This commit is a combination of 9 commits, squashed into a single commit representing the
old Vanilla Labs site hosted here.

Note: The commit messages are not very descriptive but that's how I did it in the old days, gotta live with it :P

Commits: This is a combination of 9 commits.
- test
- Replace master branch with page content via GitHub
- pushing it to our site
- cname
- v1.0
- preloader added!
- viewport
- pic for mail
- Delete CNAME